Remove newline from printf in shellquote()
It was superfluous because the previous sed command already outputs a
newline (escaped with a backslash). The space is still necessary because
without it, the shell would strip out the final newline leaving a
backslash in a potentially dangerous position. A space lets us take care
of this without needing to post-process the output (since shells ignore
extra spaces outside of quotes).