Add adb-pull-recent-photos
adb-pull-recent-photos will pull recent photos from the currently
attached android device using adb. It will default to choosing photos
taken since midnight, but a different time can be specified as the first
argument to the script.

It chooses the device with `adb -d`, and it pulls into the current
directory.