Add input hack
It's hard to get x11 to keep a consistent configuration for input
devices when reconnecting them (e.g. through USB KVM switch). This
script is a hack to reconfigure all input devices when a newly connected
input device is registered. It is certainly nowhere close to efficient,
but at least it stops my keyboard layout and mouse acceleration settings
from resetting on me.